move app layout into app component
