--- polling-app-angular-client/src/app/app-layout/app-content.component.ts
@@ -1,17 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-content',
-  template: `
-  <router-outlet></router-outlet>
-  `,
-  styles: []
-})
-export class AppContentComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
