+++ polling-app-angular-client/src/app/app.component.html
@@ -41,3 +41,8 @@
         </ng-template>
     </div>
 </nav>
+
+<div class="container">
+    <app-alert></app-alert>
+    <router-outlet></router-outlet>
+</div>
